+++ ibdl-extractors/src/websites/gelbooru/mod.rs
@@ -70,7 +70,7 @@ impl Extractor for GelbooruExtractor {
         }
     }
 
-    async fn search(&mut self, page: usize) -> Result<PostQueue, ExtractorError> {
+    async fn search(&mut self, page: u16) -> Result<PostQueue, ExtractorError> {
         let mut posts = Self::get_post_list(self, page).await?;
 
         if posts.is_empty() {
@@ -91,8 +91,8 @@ impl Extractor for GelbooruExtractor {
 
     async fn full_search(
         &mut self,
-        start_page: Option<usize>,
-        limit: Option<usize>,
+        start_page: Option<u16>,
+        limit: Option<u16>,
     ) -> Result<PostQueue, ExtractorError> {
         let blacklist = BlacklistFilter::init(
             self.active_imageboard,
@@ -103,7 +103,7 @@ impl Extractor for GelbooruExtractor {
         .await?;
 
         let mut fvec = if let Some(size) = limit {
-            Vec::with_capacity(size)
+            Vec::with_capacity(size as usize)
         } else {
             Vec::new()
         };
@@ -136,7 +136,7 @@ impl Extractor for GelbooruExtractor {
             fvec.extend(list);
 
             if let Some(num) = limit {
-                if fvec.len() >= num {
+                if fvec.len() >= num as usize {
                     break;
                 }
             }
@@ -168,7 +168,7 @@ impl Extractor for GelbooruExtractor {
         Ok(fin)
     }
 
-    async fn get_post_list(&self, page: usize) -> Result<Vec<Post>, ExtractorError> {
+    async fn get_post_list(&self, page: u16) -> Result<Vec<Post>, ExtractorError> {
         let url = format!(
             "{}&tags={}",
             self.active_imageboard.post_url(),
